+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Add saturation tools for Wazuh agent's modules ([#5697](https://github.com/wazuh/wazuh-qa/pull/5697)) \- (Framework + Tests)
 - Add Statistical Data Analyzer Module ([#5622](https://github.com/wazuh/wazuh-qa/pull/5622)) \- (Framework + Tests)
 
 ## [4.10.0] - TBD